[#14] Separate out cache tests
These all test that a response *is* cached.
