+++ cdn_acceptance_test.go
@@ -58,47 +58,6 @@ func TestRequestsGoToOriginByDefault(t *testing.T) {
 
 }
 
-// Should cache first response and return it on second request without
-// hitting origin again.
-func TestFirstResponseCached(t *testing.T) {
-	const bodyExpected = "first request"
-	const requestsExpectedCount = 1
-	requestsReceivedCount := 0
-
-	originServer.SwitchHandler(func(w http.ResponseWriter, r *http.Request) {
-		if requestsReceivedCount == 0 {
-			w.Write([]byte(bodyExpected))
-		} else {
-			w.Write([]byte("subsequent request"))
-		}
-
-		requestsReceivedCount++
-	})
-
-	url := fmt.Sprintf("https://%s/%s", *edgeHost, NewUUID())
-	req, _ := http.NewRequest("GET", url, nil)
-
-	for i := 0; i < 2; i++ {
-		resp, err := client.RoundTrip(req)
-		if err != nil {
-			t.Fatal(err)
-		}
-
-		defer resp.Body.Close()
-		body, err := ioutil.ReadAll(resp.Body)
-		if err != nil {
-			t.Fatal(err)
-		}
-		if string(body) != bodyExpected {
-			t.Errorf("Incorrect response body. Expected %q, got %q", bodyExpected, body)
-		}
-	}
-
-	if requestsReceivedCount > requestsExpectedCount {
-		t.Errorf("originServer got too many requests. Expected %d requests, got %d", requestsExpectedCount, requestsReceivedCount)
-	}
-}
-
 // Should return 403 for PURGE requests from IPs not in the whitelist. We
 // assume that this is not running from a whitelisted address.
 func TestRestrictPurgeRequests(t *testing.T) {
@@ -162,11 +121,6 @@ func TestHeaderUnspoofableClientIP(t *testing.T) {
 
 // Should not modify Host header from original request.
 func TestHeaderHostUnmodified(t *testing.T) {
-	t.Error("Not implemented")
-}
-
-// Should set a default TTL if the response doesn't set one.
-func TestDefaultTTL(t *testing.T) {
 	t.Error("Not implemented")
 }
